a horse walks into a bar and the bartender says why the long face?
